# application/provider/docs/generated/provider_swagger.yaml
basePath: /
definitions:
...
  caleta.Walletbalance200JSONResponse:
    properties:
      balance:
        description: 'Balance We use integers to represent the amount of money. To
          convert real float value to integer we multiply it by 100000. Example: $3.56
          must be represented as 356000'
        type: integer
      currency:
        allOf:
        - $ref: '#/definitions/caleta.Currency'
        description: Currency ISO 4217 currency code. Following enum contains all
          currencies supported by our system. Existing of currency in this list doesn't
          mean that all providers support native game play with it. Please contact
          us to know which provider supports which currencies.
      request_uuid:
        description: RequestUuid Standard 16-byte UUID. This id can be seen as network
          layer action. An id of an action that is generated for each our call to
          Operator. Used to sync Hub88 and Operator sides for debugging purposes.
          Operator has to respond with the same request_uuid as the one that was received
          in request.
        type: string
      status:
        allOf:
        - $ref: '#/definitions/caleta.Status'
        description: Status Response status usage:<br/>RS_OK - when transaction was
          processed correctly or was already processed before.<br/>RS_ERROR_UNKNOWN
          - use this when you need a generic error code<br/>RS_ERROR_TIMEOUT - when
          the server was disconnected by timeout<br/>RS_ERROR_INVALID_TOKEN - when
          token never existed or is not associated to this player.<br/>RS_ERROR_INVALID_GAME
          - when the game doesn't exist<br/>RS_ERROR_WRONG_CURRENCY - when the currency
          is not active or does not exist<br/>RS_ERROR_NOT_ENOUGH_MONEY - when user
          is short of money<br/>RS_ERROR_USER_DISABLED when user is banned or inactive<br/>RS_ERROR_INVALID_SIGNATURE
          - when signature doesn't match<br/>RS_ERROR_TOKEN_EXPIRED - when a new token
          exists (this rule applies only for /wallet/bet)<br/>RS_ERROR_WRONG_SYNTAX
          - message can't be decoded to JSON<br/>RS_ERROR_WRONG_TYPES - i.e. when
          a type should be integer and came as string<br/>RS_ERROR_DUPLICATE_TRANSACTION
          - when the same transaction_uuid appears but for a different user, round
          and game, otherwise returns RS_OK<br/>RS_ERROR_TRANSACTION_DOES_NOT_EXIST
          - when doing a 'win' transaction and the 'bet' transaction id doesn't exist<br/>RS_ERROR_TRANSACTION_ROLLED_BACK
          - when a 'win' transaction happens in name of a rolled back transaction
          (when receiving a 'rollback' replies with 'RS_OK'<br/>RS_ERROR_BET_LIMIT_EXCEEDED
          - when the user exceeded its limits
      user:
        description: User Unique User ID in the Operator�s system. In case of DEMO
          gameplay, this parameter may be omitted.
        type: string
    type: object
  caleta.WalletbalanceJSONRequestBody:
    properties:
      game_code:
        description: GameCode Unique Game Code on Caleta system. Can be obtained from
          `/api/game/list` endpoint. This variable intends to replace `game_id`.
        type: string
      game_id:
        description: 'GameId Deprecated: Use `game_code` instead.<br/>Unique game
          ID Caleta system. Can be obtained from `/api/game/list` endpoint.<br/>Note:
          this value may change from each server (staging, production)'
        type: integer
      request_uuid:
        description: RequestUuid Standard 16-byte UUID. This id can be seen as network
          layer action. An id of an action that is generated for each our call to
          Operator. Used to sync Hub88 and Operator sides for debugging purposes.
          Operator has to respond with the same request_uuid as the one that was received
          in request.
        type: string
      supplier_user:
        description: SupplierUser The name of the user in Provider's system (in case
          Operator needs to find user in Provder's back office or report problem with
          the user). If value is NULL, the Operator can search for their own user_id.
        type: string
      token:
        description: Token Operator's back end generates a token associated with the
          User, game, his or her current currency and maybe other parameters depending
          on the Operator's preferences. The token acts as an ID parameter for the
          game session. It is important that the currency is not changed during the
          current game session. If the User changes the currency or display unit,
          the Operator's backend needs to generate a new token and re-open the game
          with new token. In case of DEMO gameplay, this parameter may be omitted.<br/>A
          token should be valid for all bet transactions until a new one is generated,
          expired tokens should continue be valid for Win/Rollback transactions if
          is related to a previous existing bet.
        type: string
    type: object
  caleta.Walletbet200JSONResponse:
    properties:
      balance:
        description: 'Balance We use integers to represent the amount of money. To
          convert real float value to integer we multiply it by 100000. Example: $3.56
          must be represented as 356000'
        type: integer
      currency:
        allOf:
        - $ref: '#/definitions/caleta.Currency'
        description: Currency ISO 4217 currency code. Following enum contains all
          currencies supported by our system. Existing of currency in this list doesn't
          mean that all providers support native game play with it. Please contact
          us to know which provider supports which currencies.
      request_uuid:
        description: RequestUuid Standard 16-byte UUID. This id can be seen as network
          layer action. An id of an action that is generated for each our call to
          Operator. Used to sync Hub88 and Operator sides for debugging purposes.
          Operator has to respond with the same request_uuid as the one that was received
          in request.
        type: string
      status:
        allOf:
        - $ref: '#/definitions/caleta.Status'
        description: Status Response status usage:<br/>RS_OK - when transaction was
          processed correctly or was already processed before.<br/>RS_ERROR_UNKNOWN
          - use this when you need a generic error code<br/>RS_ERROR_TIMEOUT - when
          the server was disconnected by timeout<br/>RS_ERROR_INVALID_TOKEN - when
          token never existed or is not associated to this player.<br/>RS_ERROR_INVALID_GAME
          - when the game doesn't exist<br/>RS_ERROR_WRONG_CURRENCY - when the currency
          is not active or does not exist<br/>RS_ERROR_NOT_ENOUGH_MONEY - when user
          is short of money<br/>RS_ERROR_USER_DISABLED when user is banned or inactive<br/>RS_ERROR_INVALID_SIGNATURE
          - when signature doesn't match<br/>RS_ERROR_TOKEN_EXPIRED - when a new token
          exists (this rule applies only for /wallet/bet)<br/>RS_ERROR_WRONG_SYNTAX
          - message can't be decoded to JSON<br/>RS_ERROR_WRONG_TYPES - i.e. when
          a type should be integer and came as string<br/>RS_ERROR_DUPLICATE_TRANSACTION
          - when the same transaction_uuid appears but for a different user, round
          and game, otherwise returns RS_OK<br/>RS_ERROR_TRANSACTION_DOES_NOT_EXIST
          - when doing a 'win' transaction and the 'bet' transaction id doesn't exist<br/>RS_ERROR_TRANSACTION_ROLLED_BACK
          - when a 'win' transaction happens in name of a rolled back transaction
          (when receiving a 'rollback' replies with 'RS_OK'<br/>RS_ERROR_BET_LIMIT_EXCEEDED
          - when the user exceeded its limits
      user:
        description: User Unique User ID in the Operator�s system. In case of DEMO
          gameplay, this parameter may be omitted.
        type: string
    type: object
  caleta.WalletbetJSONRequestBody:
    properties:
      amount:
        description: 'Amount We use integers to represent the amount of money. To
          convert real float value to integer we multiply it by 100000. Example: $3.56
          must be represented as 356000'
        type: integer
      bet:
        description: Bet Field for metadata related to transaction, such as type of
          bet, value, time, etc. Differs from game to game. Not relevant for transaction
          processing procedure but could be useful for statistics or activity backtracking.
        type: string
      campaign_uuid:
        description: CampaignUuid Each Free Bet campaign contains an unique identifier
          on Caleta side. Campaings holds information of game where Free Bet will
          happen, total of spins, bet value and period of campaign.
        type: string
      currency:
        allOf:
        - $ref: '#/definitions/caleta.Currency'
        description: Currency ISO 4217 currency code. Following enum contains all
          currencies supported by our system. Existing of currency in this list doesn't
          mean that all providers support native game play with it. Please contact
          us to know which provider supports which currencies.
      game_code:
        description: GameCode Unique Game Code on Caleta system. Can be obtained from
          `/api/game/list` endpoint. This variable intends to replace `game_id`.
        type: string
      game_id:
        description: 'GameId Deprecated: Use `game_code` instead.<br/>Unique game
          ID Caleta system. Can be obtained from `/api/game/list` endpoint.<br/>Note:
          this value may change from each server (staging, production)'
        type: integer
      is_free:
        description: IsFree Flag which shows that transaction was generated by a promotional
          tool (FreeSpins, etc). Usually, these transactions are credited to bonus
          wallets (if available).
        type: boolean
      request_uuid:
        description: RequestUuid Standard 16-byte UUID. This id can be seen as network
          layer action. An id of an action that is generated for each our call to
          Operator. Used to sync Hub88 and Operator sides for debugging purposes.
          Operator has to respond with the same request_uuid as the one that was received
          in request.
        type: string
      reward_uuid:
        description: RewardUuid This holds the uuid of a reward given to a player
          on Caleta side. Rewards holds information about which campaigns belong,
          player, quantity left of Free Bets.
        type: string
      round:
        description: Round Game round id. Used to relate all bets and wins in one
          round. All transactions related to the same round will have the same value
          in this field. It's unique through whole system.
        type: string
      round_closed:
        description: RoundClosed Denotes when the round is closed
        type: boolean
      supplier_user:
        description: SupplierUser The name of the user in Provider's system (in case
          Operator needs to find user in Provder's back office or report problem with
          the user). If value is NULL, the Operator can search for their own user_id.
        type: string
      token:
        description: Token Operator's back end generates a token associated with the
          User, game, his or her current currency and maybe other parameters depending
          on the Operator's preferences. The token acts as an ID parameter for the
          game session. It is important that the currency is not changed during the
          current game session. If the User changes the currency or display unit,
          the Operator's backend needs to generate a new token and re-open the game
          with new token. In case of DEMO gameplay, this parameter may be omitted.<br/>A
          token should be valid for all bet transactions until a new one is generated,
          expired tokens should continue be valid for Win/Rollback transactions if
          is related to a previous existing bet.
        type: string
      transaction_uuid:
        description: TransactionUuid Unique wallet transaction.
        type: string
    type: object
  caleta.Walletcheck200JSONResponse:
    properties:
      token:
        description: Token Operator's back end generates a token associated with the
          User, game, his or her current currency and maybe other parameters depending
          on the Operator's preferences. The token acts as an ID parameter for the
          game session. It is important that the currency is not changed during the
          current game session. If the User changes the currency or display unit,
          the Operator's backend needs to generate a new token and re-open the game
          with new token. In case of DEMO gameplay, this parameter may be omitted.<br/>A
          token should be valid for all bet transactions until a new one is generated,
          expired tokens should continue be valid for Win/Rollback transactions if
          is related to a previous existing bet.
        type: string
    type: object
  caleta.WalletcheckJSONRequestBody:
    properties:
      token:
        description: Token Operator's back end generates a token associated with the
          User, game, his or her current currency and maybe other parameters depending
          on the Operator's preferences. The token acts as an ID parameter for the
          game session. It is important that the currency is not changed during the
          current game session. If the User changes the currency or display unit,
          the Operator's backend needs to generate a new token and re-open the game
          with new token. In case of DEMO gameplay, this parameter may be omitted.<br/>A
          token should be valid for all bet transactions until a new one is generated,
          expired tokens should continue be valid for Win/Rollback transactions if
          is related to a previous existing bet.
        type: string
    type: object
...
host: localhost:8083
info:
  contact: {}
  description: The exposed endpoints by the enabled game provider modules.
  title: Valkyrie Provider API
  version: '-'
paths:
  /providers/caleta/wallet/balance:
    post:
      consumes:
      - application/json
      description: Should return wallet balance for current player.
      operationId: CaletaBalance
      parameters:
      - description: Request body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/caleta.WalletbalanceJSONRequestBody'
      - description: Signature for request
        in: header
        name: X-Auth-Signature
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/caleta.Walletbalance200JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/caleta.Walletbalance200JSONResponse'
      summary: Balance
      tags:
      - Caleta
  /providers/caleta/wallet/bet:
    post:
      consumes:
      - application/json
      description: Called when the user places a bet (debit).
      operationId: CaletaBet
      parameters:
      - description: Request body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/caleta.WalletbetJSONRequestBody'
      - description: Signature for request
        in: header
        name: X-Auth-Signature
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/caleta.Walletbet200JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/caleta.Walletbet200JSONResponse'
      summary: Bet
      tags:
      - Caleta
  /providers/caleta/wallet/check:
    post:
      consumes:
      - application/json
      description: OPTIONAL - Change the initial token received for a new one that
        will be used on wallet transactions.
      operationId: CaletaCheck
      parameters:
      - description: Request body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/caleta.WalletcheckJSONRequestBody'
      - description: Signature for request
        in: header
        name: X-Auth-Signature
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/caleta.Walletcheck200JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/caleta.Walletcheck200JSONResponse'
      summary: Check
      tags:
      - Caleta
  /providers/caleta/wallet/rollback:
    post:
      consumes:
      - application/json
      description: Called when there is need to roll back the effect of the referenced
        transaction.
      operationId: CaletaRollback
      parameters:
      - description: Request body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/caleta.WalletrollbackJSONRequestBody'
      - description: Signature for request
        in: header
        name: X-Auth-Signature
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/caleta.Walletrollback200JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/caleta.Walletrollback200JSONResponse'
      summary: Rollback
      tags:
      - Caleta
  /providers/caleta/wallet/win:
    post:
      consumes:
      - application/json
      description: Called when the user wins (credit).
      operationId: CaletaWin
      parameters:
      - description: Request body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/caleta.TransactionwinJSONRequestBody'
      - description: Signature for request
        in: header
        name: X-Auth-Signature
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/caleta.Transactionwin200JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/caleta.Transactionwin200JSONResponse'
      summary: Win
      tags:
      - Caleta
...
schemes:
- httpclient
swagger: "2.0"